\name{print-methods}
\alias{print.qpcrmacro}

\title{print qpcrmacro}
\description{Prints qpcrmacro objects}
\usage{
\method{print}{qpcrmacro}(x, ...)
}
\arguments{
  \item{x}{An object of class qpcrmacro}
  \item{\dots}{\dots}
}

\author{Daniel Gerhard}

\seealso{\code{\link{qpcr}}}

\keyword{htest}
